+++ src/main/java/org/javlo/component/core/ComponentBean.java
@@ -43,6 +43,7 @@ public class ComponentBean implements Serializable, Comparable<ComponentBean> {
 	private int columnSize = -1;
 
 	private boolean repeat = false;
+	private boolean forceCachable = false;
 	private boolean modify = false;
 	private boolean nolink = false;
 	private Set<Integer> hiddenModes;
@@ -323,6 +324,14 @@ public class ComponentBean implements Serializable, Comparable<ComponentBean> {
 
 	public void setColumnSize(int columnSize) {
 		this.columnSize = columnSize;
+	}
+
+	public boolean isForceCachable() {
+		return forceCachable;
+	}
+
+	public void setForceCachable(boolean forceCachable) {
+		this.forceCachable = forceCachable;
 	}
 
 }