+++ src/main/java/org/javlo/component/links/PageReferenceComponent.java
@@ -1133,10 +1133,13 @@ public class PageReferenceComponent extends ComplexPropertiesLink implements IAc
 
 	@Override
 	public boolean isContentCachable(ContentContext ctx) {
+		if (isForceCachable()) {
+			return true;
+		}
 		if (isSessionTaxonomy(ctx)) {
 			return false;
-		}
-		return StringHelper.isTrue(getConfig(ctx).getProperty("config.cache", null), false);
+		}		
+		return StringHelper.isTrue(getConfig(ctx).getProperty("config.cache."+getCurrentRenderer(ctx), getConfig(ctx).getProperty("config.cache", null)), false);
 	}
 
 	@Override